Assign all tables in database to new database owner
This is needed in the following case:
1. Create a Postgres CR to create a database and a group role
2. Use the new database and group role (through a PostgresUser) to
create a table
3. Delete the Postgres CR which reassigns all objects owned by the group
role and the database itself to the operator's user
4. Recreate the Postgres CR

With this change all the tables created in the database are assigned
back to the group role.

+++ pkg/postgres/queries.go
@@ -3,6 +3,7 @@ package postgres
 const (
 	CREATE_DB           = `CREATE DATABASE "%s"`
 	ALTER_DB_OWNER      = `ALTER DATABASE "%s" OWNER TO "%s"`
+	ALTER_TABLE_OWNER   = `ALTER TABLE %s.%s OWNER TO "%s"`
 	CREATE_GROUP_ROLE   = `CREATE ROLE "%s"`
 	CREATE_USER_ROLE    = `CREATE ROLE "%s" WITH LOGIN PASSWORD '%s'`
 	GRANT_ROLE          = `GRANT "%s" TO "%s"`
@@ -11,4 +12,7 @@ const (
 	UPDATE_PASSWORD     = `ALTER ROLE "%s" WITH PASSWORD '%s'`
 	DROP_ROLE           = `DROP ROLE "%s"`
 	REASIGN_OBJECTS     = `REASSIGN OWNED BY "%s" TO "%s"`
+	GET_OWNED_TABLES    = `SELECT schemaname, tablename 
+		FROM pg_tables WHERE NOT schemaname in ('pg_catalog', 'information_schema')
+		AND tableowner = '%s'`
 )
